requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
yfinance>=0.2.50
pandas>=2.2.0
plotly>=5.24.0
streamlit>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.0
jugaad-data>=0.32
google-genai>=1.0.0
